+++ tests/Integration/ResourceTest.php
@@ -1,18 +1,10 @@
 <?php
 /*
- * Copyright 2023 Cloud Creativity Limited
+ * Copyright 2024 Cloud Creativity Limited
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Use of this source code is governed by an MIT-style
+ * license that can be found in the LICENSE file or at
+ * https://opensource.org/licenses/MIT.
  */
 
 declare(strict_types=1);
@@ -62,7 +54,7 @@ class ResourceTest extends TestCase
     /**
      * @return array
      */
-    public function emptyIdProvider(): array
+    public static function emptyIdProvider(): array
     {
         return [
             [''],
@@ -74,7 +66,7 @@ class ResourceTest extends TestCase
     /**
      * @return array[]
      */
-    public function createProvider(): array
+    public static function createProvider(): array
     {
         return [
             'data:required' => [
@@ -699,7 +691,7 @@ class ResourceTest extends TestCase
     /**
      * @return array
      */
-    public function updateProvider(): array
+    public static function updateProvider(): array
     {
         return [
             'data.id:required' => [
@@ -782,7 +774,7 @@ class ResourceTest extends TestCase
     /**
      * @return array
      */
-    public function emptyStringProvider(): array
+    public static function emptyStringProvider(): array
     {
         return [
             [''],
@@ -834,7 +826,7 @@ class ResourceTest extends TestCase
     /**
      * @return array
      */
-    public function nonObjectProvider(): array
+    public static function nonObjectProvider(): array
     {
         return [
             ['true'],